requirements.txt
Pillow
smartcrop
vk_getter
python-dotenv
aiogram
aiogram_dialog
